Ensure smooth deployment
Many parts of the code were altered as there turned out to be many bugs in the code. Both
frontend and backend code had to be altered.

The changes include subtle changes to the User model, changes to API call responses, changes to services and components, as well as the main.ts and app components.